Update README.md
# Chi-Milestone
New website
Things I would like the website or sites to do is to be able to provide online learning and live classes. My other idea is to create a website/program that can track individuals (or groups) health parameters, along with providing exercise and healthy lifestyle prescriptions for users. My initial goal is to achieve building a basic website, then develop my skills to build a more advanced website. When I have achieved this, I will have to develop the skills to be a top code writer for this industry and will be able to provide my skills to other industries.  


Started some basic HTML and styling for my index page 25/11/20